better handling for stuck posts and retries
